feat: fixed compiler errors for fb150

+++ cecs_math/cecs_math/arithmetic/cecs_integer_arithmetic.c
@@ -100,8 +100,8 @@ extern inline uint8_t cecs_gather_lsb8_u8(const uint64_t vec);
 extern inline uint8_t cecs_gather_msn8_u4(const uint32_t vec);
 extern inline uint8_t cecs_gather_lsn8_u4(const uint32_t vec);
 
-extern inline uint8_t cecs_gather_msnh15_u4(const uint64_t vec);
-extern inline uint8_t cecs_gather_lsnh15_u4(const uint64_t vec);
+extern inline uint16_t cecs_gather_msnh15_u4(const uint64_t vec);
+extern inline uint16_t cecs_gather_lsnh15_u4(const uint64_t vec);
 
 extern inline uint64_t cecs_scatter_msb8_u1(const uint8_t vec);
 extern inline uint64_t cecs_scatter_lsb8_u1(const uint8_t vec);
